Move filesize fmting logic to new `value/filesize`
This is a `git blame` tollbooth! Move along!
